Remove unify_direct.rs
We don't generate new constraints during
unification for collections, instead, we
generate constraint for all set items
within the generation stage.

unify_direct.rs was only necessary for
unifying newly generated constraints.

- Remove parents from primitives, which
  caused some errors during the type
  checking of arithmetic operations.

+++ src/check/resource/primitive/int.py
@@ -1,4 +1,6 @@
-class int(float):
+from typing import Union
+
+class int:
     def __init__(self, arg: int): pass
 
     def __add__(self, other: int) -> int: pass
